+++ NEWS.md
@@ -6,6 +6,7 @@
 * [MODSOURCE-930](https://folio-org.atlassian.net/browse/MODSOURCE-930) Cannot update MARC authority record by created user
 * [MODSOURCE-962](https://folio-org.atlassian.net/browse/MODSOURCE-962) MARC Modifications fail for Updates
 * [MODSOURCE-964](https://folio-org.atlassian.net/browse/MODSOURCE-964) 003 is being deleted when overlaying a MARC authority record with Data Import
+* [MODSOURCE-936](https://folio-org.atlassian.net/browse/MODSOURCE-936) Upgrade module to Vert.x 5.0
 
 ## 2025-03-13 5.10.0
 * [MODSOURMAN-1278](https://folio-org.atlassian.net/browse/MODSOURMAN-1278) Data import job stuck in progress with 99-100%